fix(service-worker): configure service worker for better caching and navigation
#Caution with PWA
Be careful if you are using Angular SSR in combination with the Angular PWA service worker because the behavior deviates from default SSR. The initial request will be server-side rendered as expected. However, subsequent requests are handled by the service worker and thus client-side rendered.

Most of the time that's what you want. Nevertheless, if you want a fresh request you can use the freshness option as Angular PWA navigationRequestStrategy. This approach will try a network request and fall back to the cached version of index.html when offline.

- Configured the service worker to prefetch essential resources for faster loading.
- Added `updateMode: prefetch` to ensure resources are updated in the background.
- Included `/assets/**` in the app resource files to cache assets.
- Added `eot` to the list of cacheable files.
- Set `navigationRequestStrategy` to `freshness` to serve the latest content for navigation requests.